+++ srcs/ft_setcursor.c
@@ -6,7 +6,7 @@
 /*   By: mrantil <mrantil@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/20 13:43:20 by mrantil           #+#    #+#             */
-/*   Updated: 2022/11/07 16:08:51 by mrantil          ###   ########.fr       */
+/*   Updated: 2022/11/11 15:59:40 by mrantil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@
  * @param col The column number to move the cursor to.
  * @param row the row number of the cursor
  */
-void	ft_setcursor(int col, int row)
+void	ft_setcursor(ssize_t col, ssize_t row)
 {
 	char	*cm;
 
 	cm = tgetstr("cm", NULL);
-	tputs(tgoto(cm, col, row), 1, ft_putc);
+	tputs(tgoto(cm, (int)col, (int)row), 1, ft_putc);
 }
